Split the XML-to-HTML converter into components
The converter now provides a set of functions for converting XML into
HTML, and a set of web page components for organizing the converted
HTML into HTML documents.

The intention is that this makes the converter more general purpose: you
can customize the produced documentation by introducing your own
replacement components, instead of having to modify the converter code
directly.